Q's practice 2: Convolutional T test
This practice simulated two datasets. In the 1st data set, the
activations are more clustered, while the activations are more isolated
in the 2nd data set.

It turns out that by convolving the data, T test is more able to detect
the signal in the 1st data set, because convolutional T test is able to
take account for the signals from neighboring units.