This is an attempt for cross validation
CV was implemented. It fits Lasso on the training set, and test these
weights on the testing set.

In the end, it will generate a plot showing the accuracy. It turns out
that the best lambda for the training set is usually not the best
lambda for the testing set.

Next: I plan to finish the entire cross validation.